Updates dependencies, PHP, and plugin list
** Why are these changes being introduced:

* It has been a while since we've updated the WordPress application and
  tooling behind the Hayden Renovation site. This feels like a good time
  to catch up on these changes.

** Relevant ticket(s):

* n/a

** How does this address that need:

This does a few things:

* Updates PHP from the 7.3 branch to 7.4. This is not the current
  release, but is still supported by the community. The 7.3 branch has
  reached its end of life, and 8.x is not available on Pantheon that I
  can tell.
* Updates site dependencies through running `composer update`
* Updates a few WordPress plugins, distinct from what Composer does on
  its own.
* Removes two unused plugins (HSTS and Jetpack). These were either
  never activated (Jetpack) or are now handled at the platform level
  (HSTS)

What this plugin does not do - yet? - is to enable the Sentry plugin.
The plugin is available, but not enabled via the configuration bundle.
This is supposed to be possible, but thus far the ability to export the
current configuration back to source control is not working. This is the
topic I'm looking for help on tomorrow.

** Document any side effects to this change:

In its current state, the PR should have no side effects - only updates
and the removal of two unused plugins. If we can get the config bundle
updated, this PR would enable that plugin.